refactor: promote undo.rs to undo/ directory with snapshot + rollback split
Split the two independent undo subsystems into separate files:
- undo/snapshot.rs: Persistent disk-based UndoSnapshot for `homeboy undo`
- undo/rollback.rs: Ephemeral InMemoryRollback for per-chunk fixer verification
- undo/mod.rs: Re-exports everything to preserve existing import paths

872 lines → two focused ~400-line files. Zero behavior change.
